# [read-unfinalized-object] [part1] Relax file-cache checks for unfinalized objects