--- a/internal/cache/file/cache_handle.go
+++ b/internal/cache/file/cache_handle.go
@@ -167,6 +167,9 @@ func (fch *CacheHandle) Read(ctx context.Context, bucket gcs.Bucket, object *gcs
 	// If the end-offset is within the cached size, then the read will be served from the cache.
 	// If offset is within the cached size and end-offset is beyond the cached-size, then
 	// we will fallback to GCS in the later logic, but we should still create the cache download job even in that case.
+	//
+	// Note: Change the below check to `(offset + len(dst)) > int64(fileInfoData.FileSize))` if the below
+	// check causes a problem in any edge-case.
 	if bucket.BucketType().Zonal && object.IsUnfinalized() && offset >= int64(fileInfoData.FileSize) {
 		err = util.ErrFallbackToGCS
 		return