### Description
The change is based on [design doc](https://docs.google.com/document/d/1eKAsVbvmiBYtPs8h3jbapN8jP_Pmegb4_AW8WlTdNts/edit?usp=sharing&resourcekey=0-z2SvePxO8kTVnL64-qHq5w)

Changes in behavior for unfinalized objects:
- Don't invalidate file-cache entry if size of completed download entry >= intended downloaded size
- Don't fall back to GCS from file-cache read if size of completed download entry >= intended upper limit of offset to be read.
- Fallback to GCS while reading from cache_handle if requested read isn't available from the cached file-info data size. If partial read is possible from the cache, then go ahead and start the download job.

Additional changes
- Added a simple utility for easily checking if an object is unfinalized.
- Enable zonal bucket unit tests for TestFileCacheReader, TestRandomReader and TestReadManager.
- Add new unit tests corresponding to the above code changes.
- Update existing unit tests corresponding to the above code changes.

~Dependencies~
- ~This builds on top of #3387 and is thus blocked for merge until that PR is merged~

Out of scope:
- This does not add support for reading unfinalized objects itself. That's covered in the in https://github.com/GoogleCloudPlatform/gcsfuse/pull/3405 .
- The changes in this don't touch any non-zonal non-unfinalized objects in any way.

### Link to the issue in case of a bug fix.
[b/421294827](http://b/421294827)

### Testing details
1. Manual - Tested manually.
2. Unit tests - Tested automatically by this PR. Add more UTs for the changes in the code.
3. Integration tests - ran with presubmit ([only zonal](https://fusion2.corp.google.com/ci/kokoro/prod:gcsfuse%2Fgcp_ubuntu%2Fpresubmits%2Fperf_tests%2Fpresubmit/activity/b21fc64b-349f-4287-bff6-2bc06371fc02/details), [e2e zonal, non-zonal and perf test run](http://fusion2/ci/kokoro/prod%3Agcsfuse%2Fgcp_ubuntu%2Fpresubmits%2Fperf_tests%2Fpresubmit/activity/fa4f953d-79d2-445e-93a7-3306010e4d72))
4. Perf tests results
```md
+--------+------------+--------------+------------+--------------+--------------+
| Branch | File Size  |   Read BW    |  Write BW  | RandRead BW  | RandWrite BW |
+--------+------------+--------------+------------+--------------+--------------+
| Master |  0.25MiB   | 555.31MiB/s  |  1.2MiB/s  |  77.32MiB/s  |  1.17MiB/s   |
|   PR   |  0.25MiB   | 581.98MiB/s  | 1.21MiB/s  |  77.26MiB/s  |   1.2MiB/s   |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 48.828MiB  | 3900.45MiB/s | 85.61MiB/s | 1554.19MiB/s |  84.39MiB/s  |
|   PR   | 48.828MiB  | 3849.86MiB/s | 85.42MiB/s | 1499.03MiB/s |  84.85MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 976.562MiB | 3742.3MiB/s  | 79.22MiB/s | 763.35MiB/s  |  39.24MiB/s  |
|   PR   | 976.562MiB | 3711.34MiB/s | 77.93MiB/s | 1296.53MiB/s |  38.71MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
+--------+------------+--------------+------------+--------------+--------------+
```

### Any backward incompatible change? If so, please explain.